spinning-square.sky doesn't spin
The semantics of requestAnimationFrame and scheduleVisualUpdate are slightly
different. If you're inside a requestAnimationFrame, then calling
requestAnimationFrame should schedule you for next frame. However, if you're
inside a requestAnimationFrame and you scheduleVisualUpdate, that's a noop
because we've already got a visual update pending. When we merge the codepaths,
we broke requestAnimationFrame because it became a noop when called inside a
request animation frame.

We really ought to split up the two roles of requestAnimationFrame (run before
the frame boundary and run once per frame).

R=esprehn@chromium.org, eseidel@chromium.org

Review URL: https://codereview.chromium.org/880013003