Add Tonic macros for dispatching to static methods
DartArgIterator was starting at argument index 1 because it assumed that the
arguments include a pointer to the "this" object for an instance method.
DART_REGISTER_NATIVE makes the same assumption in the argument count.

This change adds versions of the macros that are suitable for static methods
with no implicit arguments.