[Impeller] Ignore opt-outs on iOS devices. (#55808)
Addresses https://github.com/flutter/flutter/issues/155541

This does **not** remove Skia itself from the build. I'll stage that in a followup. Want to keep the scope of this patch small.

This is the new behavior:

* Impeller is the default on iOS devices **and** simulators.
* On iOS devices **only**, all flags and plist options are ignored.
* On iOS simulators **only**, Flutter will used Impeller's Metal backend by default and fallback to the Null backend if Metal device access is not available.
* On iOS simulators **only**, Flutter can pick Skia using the command line flags or plist flags. This is to allow users one more month to migrate as communicated by @zanderso.

The settings objects `enable_impeller` field is now `static constexpr` to avoid accidentally disabling the flag while it still exists.